requirements.txt
Django
ussd-framework
djangorestframework
markdown